+++ src/fontra/client/core/context-menu.js
@@ -1,4 +1,4 @@
-import { reversed, buildShortcut } from "./utils.js";
+import { reversed } from "./utils.js";
 
 export const MenuItemDivider = { title: "-" };
 
@@ -136,9 +136,29 @@ export class ContextMenu {
     title.innerText = typeof item.title === "function" ? item.title() : item.title;
 
     const shortCut = document.createElement("span");
-    shortCut.innerHTML = buildShortcut(item.shortCut);
+    shortCut.innerHTML = this.buildShortCutString(item.shortCut);
     titleWrapper.append(title, shortCut);
     return titleWrapper;
+  }
+
+  buildShortCutString(shortCutDefinition) {
+    let shorcutCommand = "";
+
+    if (shortCutDefinition) {
+      const isMac = navigator.platform.toLowerCase().indexOf("mac") >= 0;
+
+      if (shortCutDefinition.shiftKey) {
+        shorcutCommand += isMac ? "&#8679;" : "Shift+"; // ⇧ or Shift
+      }
+      if (shortCutDefinition.metaKey) {
+        shorcutCommand += isMac ? "&#8984;" : "Ctrl+"; // ⌘ or Ctrl
+      }
+      if (shortCutDefinition.keysOrCodes) {
+        shorcutCommand += shortCutDefinition.keysOrCodes.toUpperCase();
+      }
+    }
+
+    return shorcutCommand;
   }
 }
