+++ src/fontra/client/css/context-menu.css
@@ -86,7 +86,7 @@
 }
 
 .context-menu-item {
-  padding: 0.2em 1.5em 0.2em 1em; /* top, right, bottom, left */
+  padding: 0.2em 0.8em 0.2em 1em; /* top, right, bottom, left */
   color: #808080a0;
 }
 
@@ -102,5 +102,6 @@
 
 .context-menu-item > div {
   display: flex;
+  gap: 0.5em;
   justify-content: space-between;
 }
